backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
whitenoise>=6.6
gunicorn>=21.0
dj-database-url>=2.1
